afxdp/mpsc_inbox: relocate inline tests to mpsc_inbox_tests.rs
Same playbook. 181-LOC inline test block extracted. mpsc_inbox.rs: 367 → 189 LOC.
cargo build/test 92/878.
